fix bug role name ga muncul

+++ resources/views/roles/index.blade.php
@@ -7,10 +7,9 @@
                 <h2>Role Management</h2>
             </div>
             <div class="pull-right">
-                {{-- @can('role-create')
+                @can('role-create')
                     <a class="btn btn-success" href="{{ route('roles.create') }}"> Create New Role</a>
-                @endcan --}}
-                <a class="btn btn-success" href="{{ route('roles.create') }}"> Create New Role</a>
+                @endcan
             </div>
         </div>
     </div>
@@ -37,17 +36,14 @@
                 <td>
                     <a class="btn btn-info" href="{{ route('roles.show', $role->id) }}">Show</a>
                     <a class="btn btn-primary" href="{{ route('roles.edit', $role->id) }}">Edit</a>
-                    {{-- @can('role-edit')
+                    @can('role-edit')
                         <a class="btn btn-primary" href="{{ route('roles.edit', $role->id) }}">Edit</a>
-                    @endcan --}}
-                    {{-- @can('role-delete')
+                    @endcan
+                    @can('role-delete')
                         {!! Form::open(['method' => 'DELETE', 'route' => ['roles.destroy', $role->id], 'style' => 'display:inline']) !!}
                         {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
                         {!! Form::close() !!}
-                    @endcan --}}
-                    {!! Form::open(['method' => 'DELETE', 'route' => ['roles.destroy', $role->id], 'style' => 'display:inline']) !!}
-                    {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
-                    {!! Form::close() !!}
+                    @endcan
                 </td>
             </tr>
         @endforeach
